+++ database/migrations/2020_06_24_172207_bg_betting.php
@@ -20,9 +20,9 @@ class BgBetting extends Migration
             $table->string('login_ip',30)->default('')->comment('登陆ip');
             $table->string('game_name',50)->default('')->comment('游戏名称');
             $table->string('room_num',32)->default('')->comment('游戏房间');
-            $table->integer('bottom_pour')->default(0)->comment('下注金额');
-            $table->integer('group_money')->default(0)->comment('派彩金额');
-            $table->string('bottom_pour_time',50)->default('')->comment('下注时间');       
+            $table->decimal('bottom_pour',9,2)->default(0)->comment('下注金额');
+            $table->decimal('group_money',9,2)->default(0)->comment('派彩金额');
+            $table->string('bottom_pour_time',50)->default('')->comment('下注时间');
             $table->timestamps();
         });
     }
